# Write cache files to disk asynchronously.