Judging from the results of https://github.com/jdalton/reify-perf-check, this noticeably reduces the overhead from caching.

Before:
```sh
~/dev/reify-perf-check% npm test

> perf-example@0.1.0 test /Users/ben/dev/reify-perf-check
> node index

NO REIFY CACHE
loading cjs: 204.566ms
loading reify: 46.905ms
loading cjs after reify: 149.188ms
loading esm: 835.038ms
loading esm.gz: 799.494ms

REIFY CACHED
loading cjs: 110.203ms
loading esm: 248.458ms
loading esm.gz: 330.240ms
```

After:
```sh
~/dev/reify-perf-check% npm test

> perf-example@0.1.0 test /Users/ben/dev/reify-perf-check
> node index

NO REIFY CACHE
loading cjs: 196.121ms
loading reify: 44.745ms
loading cjs after reify: 142.679ms
loading esm: 709.120ms
loading esm.gz: 700.499ms

REIFY CACHED
loading cjs: 78.973ms
loading esm: 240.221ms
loading esm.gz: 331.046ms
```