@jdalton 2017-03-21
Ah nice! I dig that revisiting smth once other optimizations are in place shows the impact more. Woot! Btw the minified+zopflied bundle is down to 30 kB.