@jdalton 2017-03-29
I think once this PR is rebased it's fine to go ahead and test the waters with it. Worse case we adjust a bit for edge cases.